Revert "Merge branch 'UpdateRNGseeding' into 'master'"

See merge request atlas/athena!49694

+++ Control/CxxUtils/CxxUtils/crc64.h
@@ -155,7 +155,7 @@ uint64_t crc64 (const std::string& s);
  * @param x The integer to add.
  * @return The new CRC.
  */
-uint64_t crc64addint (uint64_t crc, uint64_t x);
+uint64_t crc64addint (uint64_t crc, unsigned int x);
 
 
 /**
